[android] rework android start to accept context instead of Application. Also reworked to be able to use startKoin from elsewhere than Application class.
